requirements.txt
python-dotenv==1.0.1
aiogram==3.5.0
aiohttp==3.9.5
aiocache==0.12.2